+++ app/Models/Feedback_template.php
@@ -2,25 +2,11 @@
 
 namespace App\Models;
 
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Relations\{HasOne, HasMany};
-
-class Feedback_template extends Model
+/**
+ * @deprecated Use FeedbackTemplate instead
+ */
+class Feedback_template extends FeedbackTemplate
 {
-    use HasFactory;
-
-    protected $fillable = [
-        'name',
-        'title'
-    ];
-    public function feedbacks(): HasMany
-    {
-        return $this->hasMany(Feedback::class);
-    }
-
-    public function questions(): HasMany
-    {
-        return $this->hasMany(Question::class);
-    }
+    // This class extends FeedbackTemplate to maintain backward compatibility
+    // while following Laravel naming conventions
 }
