+++ app/Models/Question_template.php
@@ -2,19 +2,11 @@
 
 namespace App\Models;
 
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Relations\HasMany;
-
-class Question_template extends Model
+/**
+ * @deprecated Use QuestionTemplate instead
+ */
+class Question_template extends QuestionTemplate
 {
-    protected $fillable = [
-        'type',
-        'min_value',
-        'max_value'
-    ];
-
-    public function questions(): HasMany
-    {
-        return $this->hasMany(Question::class);
-    }
+    // This class extends QuestionTemplate to maintain backward compatibility
+    // while following Laravel naming conventions
 }
