+++ refactoring-checklist.md
@@ -1,0 +1,66 @@
+# Feedback Forms Refactoring Checklist
+
+This document outlines recommended improvements to enhance code quality, maintainability, and adherence to best practices in the Feedback Forms application.
+
+## Database & Model Architecture
+
+- [x] **Standardize model naming conventions** - Models like `Feedback_template` use snake_case instead of Laravel's conventional PascalCase (e.g., `FeedbackTemplate`)
+- [ ] **Evaluate the `already_answered` fieimage.pngld** - This field appears to be redundant as it's calculated dynamically via `getSubmissionCountAttribute()`
+- [ ] **Consolidate duplicate query logic** - Several repeated database queries exist in `Feedback` model methods (`getSubmissionCountAttribute`, `getUniqueSubmissionsCount`, etc.)
+- [ ] **Review relationship naming conventions** - Some relationship methods use snake_case (e.g., `feedback_template()`) instead of camelCase
+- [ ] **Consider implementing soft deletes** - Missing for survey-related models where data preservation might be important
+
+## Service Layer
+
+- [ ] **Break down `SurveyService`** - This class exceeds 1600 lines and handles too many responsibilities, violating the Single Responsibility Principle
+- [ ] **Implement template strategy pattern** - Template-specific logic is scattered with conditionals; could be refactored into dedicated strategy classes
+- [ ] **Extract statistics calculation** - Move statistics logic from `SurveyService` to a dedicated `StatisticsService`
+- [ ] **Reduce method sizes** - Methods like `storeResponses()` and `storeTemplateSpecificResponses()` are excessively long and complex
+- [ ] **Standardize error handling** - Inconsistent error handling approaches exist across service methods
+
+## View Layer
+
+- [ ] **Extract business logic from Blade templates** - `statistics.blade.php` contains complex PHP logic that should be moved to controllers/services
+- [ ] **Convert repeated template structures to components** - Many templates have similar patterns that could be reused via components
+- [ ] **Refactor Alpine.js usage** - Consider organizing Alpine.js code into dedicated JavaScript files for larger interactive components
+- [ ] **Implement view models** - Use view models or data transfer objects to reduce complex data transformation in templates
+
+## Code Organization & Architecture
+
+- [ ] **Consistent repository pattern usage** - Repository pattern is used for some models but not others
+- [ ] **Improve error logging** - Current error logging includes raw exceptions that could be better structured and categorized
+- [ ] **Extract form request validation** - Move complex validation logic from services/controllers to dedicated form request classes
+- [ ] **Enhance API documentation** - PHPDoc comments are inconsistent across the codebase
+
+## Performance Considerations
+
+- [ ] **Review N+1 query issues** - Check for potential N+1 query problems, especially in statistics generation
+- [ ] **Implement caching** - Consider caching for frequently accessed statistics data
+- [ ] **Query optimization** - Review complex queries in `SurveyService` and `Feedback` model for optimization opportunities
+- [ ] **Index usage evaluation** - Review if all required indexes are properly defined for performance-critical queries
+
+## Code Quality & Standards
+
+- [ ] **Internationalize hardcoded strings** - Various hardcoded strings throughout the codebase should use Laravel's localization
+- [ ] **Remove commented-out code** - Clean up commented code and "TODO" items that are no longer relevant
+- [ ] **Standardize method return type declarations** - Many methods lack proper return type declarations
+- [ ] **Improve variable naming** - Some variables have generic names (`$data`, `$result`, etc.) that could be more descriptive
+
+## Security Considerations
+
+- [ ] **Review authorization controls** - Ensure proper authorization checks are in place for all survey operations
+- [ ] **Validate survey access mechanisms** - Review the access key generation and validation logic for potential issues
+- [ ] **Input validation** - Enhance input validation for user-submitted survey responses
+
+## Testing Improvements
+
+- [ ] **Increase test coverage** - Add or enhance tests for critical components, especially around survey submission and statistics
+- [ ] **Implement integration tests** - Focus on testing the interaction between components
+- [ ] **Create test factories** - Ensure all models have appropriate factory definitions for testing
+
+## Technical Debt
+
+- [ ] **Address deprecated functionality** - Legacy functionality and conditionals marked with comments like "can be removed after frontend updates"
+- [ ] **Review database migration sequence** - Consider consolidating incremental migrations that modify the same tables
+- [ ] **Document complex algorithms** - Improve documentation for complex calculations in `SurveyAggregationService`
+- [ ] **Fix inconsistent coding styles** - Ensure adherence to a consistent coding style throughout the codebase
